% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{barPlotScale}
\alias{barPlotScale}
\title{Select a suitable color palette for time series bar plots}
\usage{
barPlotScale(pltdata, subcatvar)
}
\arguments{
\item{pltdata}{The data getting plotted}

\item{subcatvar}{The subcategory selected}
}
\description{
Select a suitable color palette for time series bar plots
}
\keyword{internal}
