% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{barPlotTheme}
\alias{barPlotTheme}
\title{The theme for bar plots on the dashboard}
\usage{
barPlotTheme(legendPos = "right")
}
\arguments{
\item{legendPos}{The position of the legend}
}
\description{
The theme for bar plots on the dashboard
}
\keyword{internal}
