% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{cleanPlotData}
\alias{cleanPlotData}
\title{Filter out data that cannot be plotted}
\usage{
cleanPlotData(plotData, startYear = NULL, endYear = NULL)
}
\arguments{
\item{plotData}{Data frame containing the data for a plot.}

\item{startYear}{Minimum year to display. If NULL, searches year column for
minimum in data.}

\item{endYear}{Maximum year to display. If NULL, searches year column for
maximum in data.}
}
\value{
Cleaned data frame.
}
\description{
Filter out data that cannot be plotted
}
