% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{convertProjectToLongform}
\alias{convertProjectToLongform}
\title{Convert all queries in the project from wideform to long form}
\usage{
convertProjectToLongform(projData)
}
\arguments{
\item{projData}{Project data to convert}
}
\description{
Assumes that a wideform table has no 'year' column and several Xyear columns.
}
