% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{convertToWideform}
\alias{convertToWideform}
\title{Convert a data frame to wideform}
\usage{
convertToWideform(data)
}
\arguments{
\item{data}{Data frame to convert, likely the output of a query.}
}
\description{
Takes a data frame with a 'year' column and creates a new column for each
distinct year.
}
