% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{countUniqueSubcatValues}
\alias{countUniqueSubcatValues}
\title{Counts the number of unique values for a query subcategory}
\usage{
countUniqueSubcatValues(prj, scenario, query)
}
\arguments{
\item{prj}{Project data structure.}

\item{scenario}{Name of the scenario.}

\item{query}{Name of the query.}
}
\description{
Counts the number of unique values for a query subcategory
}
