% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{default.plot}
\alias{default.plot}
\title{Plot a default panel}
\usage{
default.plot(label.text = "No data selected")
}
\arguments{
\item{label.text}{Text to display in the middle of the panel}
}
\description{
Mainly intended for use when no data has been loaded.
}
