% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{determineMapset}
\alias{determineMapset}
\title{Figure out which map to plot a query on.}
\usage{
determineMapset(prjdata, pltscen, query)
}
\arguments{
\item{prjdata}{Project data structure}

\item{pltscen}{Name of the scenario to plot}

\item{query}{Name of the GCAM query to plot}
}
\value{
Base map suitable for plotting this data
}
\description{
Right now we assume that if the query table contains a 'basin' column, then
we want to plot on the basin map; otherwise we plot on the region map.
}
