% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getMapLimits}
\alias{getMapLimits}
\title{Select suitable scale limits for a plot}
\usage{
getMapLimits(pltdata, is.diff)
}
\arguments{
\item{pltdata}{The data being plotted}

\item{is.diff}{Boolean indicating whether the plot is a difference plot}
}
\description{
Select suitable scale limits for a plot
}
\keyword{internal}
