% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getMapPalette}
\alias{getMapPalette}
\title{Select a suitable color palette}
\usage{
getMapPalette(is.diff)
}
\arguments{
\item{is.diff}{Boolean indicating whether the plot is a difference plot}
}
\description{
Select a suitable color palette
}
\keyword{internal}
