% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getMapParams}
\alias{getMapParams}
\title{Get projection parameters for the pre-defined projections}
\usage{
getMapParams(projselect)
}
\arguments{
\item{projselect}{Name of the predefined projection}
}
\description{
Valid inputs are "global", "lac" (Latin America and Caribbean), "usa",
"china", and "africa".
}
\keyword{internal}
