% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getNewSubcategory}
\alias{getNewSubcategory}
\title{Get the subcategory to show for a new query}
\usage{
getNewSubcategory(prj, scenario, query, oldSubcategory)
}
\arguments{
\item{prj}{Project data structure.}

\item{scenario}{Name of the scenario.}

\item{query}{Name of the query.}

\item{oldSubcategory}{The subcategory for the previous query.}
}
\description{
A subcategory is defined as column that doesn't define the scenario, year,
units, or value. When a new query is selected, the new subcategory should be
the same as one in the previous query if it exists. If the new query does not
have that subcategory, a new one is chosen based on number of unique values.
If the subcategory contains only one unique element, it is not included.
}
