% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getProjectName}
\alias{getProjectName}
\title{Get the name of the project for display}
\usage{
getProjectName(rFileinfo)
}
\arguments{
\item{rFileinfo}{Reactive fileinfo object returned by the file browser in the UI}
}
\description{
Returns a place holder string if no project has been loaded yet.
}
