% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getProjectScenarios}
\alias{getProjectScenarios}
\title{Get the scenarios in the project for display}
\usage{
getProjectScenarios(rFileinfo, concat = NULL)
}
\arguments{
\item{rFileinfo}{Reactive fileinfo object returned by file browser in the UI.}

\item{concat}{Separator string to use when concatenating scenario names.}
}
\description{
Returns a place holder string if no project has been loaded yet.
}
