% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getQuerySubcategories}
\alias{getQuerySubcategories}
\title{Get the subcategories for a query}
\usage{
getQuerySubcategories(prj, scenario, query)
}
\arguments{
\item{prj}{Project data structure.}

\item{scenario}{Name of the scenario.}

\item{query}{Name of the query.}
}
\description{
A subcategory is defined as column that doesn't define the scenario, year,
units, or value. If the subcategory contains only one unique element, it is
not included.
}
