% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getQueryYears}
\alias{getQueryYears}
\title{Get the years for which a query is defined}
\usage{
getQueryYears(prj, scenario, query)
}
\arguments{
\item{prj}{Project data structure}

\item{scenario}{Name of the scenario}

\item{query}{Name of the query}
}
\description{
Get the years for which a query is defined
}
