% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getTimePlotData}
\alias{getTimePlotData}
\title{Get the data frame of the current time plot being shown}
\usage{
getTimePlotData()
}
\value{
A data frame containing the values currently being plotted
}
\description{
Get the data frame of the current time plot being shown
}
