% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{isGrid}
\alias{isGrid}
\title{Indicate whether a query is a gridded data set}
\usage{
isGrid(prj, scenario, query)
}
\arguments{
\item{prj}{Project data structure}

\item{scenario}{Name of the scenario}

\item{query}{Name of the query}
}
\description{
Indicate whether a query is a gridded data set
}
