% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{plotScenComparison}
\alias{plotScenComparison}
\title{Plot values over time as a bar chart}
\usage{
plotScenComparison(prjdata, query, scens, diffscen, subcatvar, rgns)
}
\arguments{
\item{prjdata}{A project data structure}

\item{query}{Name of the query to plot}

\item{scens}{List of scenario names to plot}

\item{diffscen}{Name of the difference scenario, or NULL if none}

\item{subcatvar}{Variable to use for subcategories in the plot}

\item{rgns}{Regions to filter to}
}
\description{
Plot values over time as a bar chart
}
