% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{plotTime}
\alias{plotTime}
\title{Plot values over time as a bar chart}
\usage{
plotTime(prjdata, query, scen, diffscen, subcatvar, filter, rgns)
}
\arguments{
\item{prjdata}{A project data structure}

\item{query}{Name of the query to plot}

\item{scen}{Name of the scenario to plot}

\item{diffscen}{Name of the difference scenario, or NULL if none}

\item{subcatvar}{Variable to use for subcategories in the plot}

\item{filter}{If TRUE, then filter to regions in the rgns argument}

\item{rgns}{Regions to filter to, if filter is TRUE.}
}
\description{
Plot values over time as a bar chart
}
