% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\docType{data}
\name{rgnlists}
\alias{rgnlists}
\alias{africa.rgns}
\alias{north.america.rgns}
\alias{lac.rgns}
\alias{europe.rgns}
\alias{asiapac.rgns}
\title{Region lists by continent}
\format{An object of class \code{character} of length 5.}
\usage{
africa.rgns

north.america.rgns

lac.rgns

europe.rgns

asiapac.rgns
}
\description{
Lists of GCAM regions grouped by continent.
}
\keyword{datasets}
