% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{summarize.unit}
\alias{summarize.unit}
\title{Summarize the unit column of a GCAM data frame by taking the most common
entry.}
\usage{
summarize.unit(unitcol)
}
\arguments{
\item{unitcol}{Character vector of unit names.}
}
\description{
In theory the unit should have a single, common value, but in practice GCAM
isn't always great about getting its unit strings consistent.
}
\keyword{internal}
