% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{uiStateValid}
\alias{uiStateValid}
\title{Indicate whether the UI is in an obviously invalid state.}
\usage{
uiStateValid(prj, scenario, query)
}
\arguments{
\item{prj}{Project data structure}

\item{scenario}{Scenario name}

\item{query}{Query name}
}
\value{
Boolean indicating whether the UI state appears to be valid.
}
\description{
Invalid states frequently occur as transients when a new project is being
loaded and the UI elements are being updated.
}
