% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{updateRegionFilter}
\alias{updateRegionFilter}
\title{Update the checkbox filters when select/deselect all button is pressed}
\usage{
updateRegionFilter(session, btnId, groupId, selectAll, choices)
}
\arguments{
\item{session}{The main session.}

\item{btnId}{The id of the actionButton that was pressed.}

\item{groupId}{The id of the checkboxGroupInput to act on.}

\item{selectAll}{If TRUE, select all checkboxes in the group defined by
groupId. If not TRUE then deselect.}

\item{choices}{The labels of the checkboxes.}
}
\description{
Update the checkbox filters when select/deselect all button is pressed
}
