% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\docType{data}
\name{year.regex}
\alias{year.regex}
\title{Regular expression for year columns}
\format{An object of class \code{character} of length 1.}
\usage{
year.regex
}
\description{
Regular expression for year columns
}
\keyword{datasets}
