% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{cleanPlotData}
\alias{cleanPlotData}
\title{Filter out data that cannot be plotted}
\usage{
cleanPlotData(plotData)
}
\arguments{
\item{plotData}{Data frame containing the data for a plot.}
}
\value{
Cleaned data frame.
}
\description{
Filter out data that cannot be plotted
}
