% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getPlotData}
\alias{getPlotData}
\title{Extract and format data for a plot}
\usage{
getPlotData(prjdata, query, pltscen, diffscen, key, filtervar = NULL,
  filterset = NULL)
}
\arguments{
\item{prjdata}{Project data structure}

\item{query}{Name of the query to plot}

\item{pltscen}{Name of the scenario to plot}

\item{key}{Aggregation variable.  (e.g., 'region' or 'sector')}

\item{filtervar}{If not NULL, filter on this variable before aggregating}

\item{diffscenDifference}{scenario, if any}

\item{filterset:}{Set of values to include in the filter operation.  Ignored
if filtervar is NULL.}
}
\description{
Extract and format data for a plot
}
\keyword{internal}
