% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{getScenarioQueries}
\alias{getScenarioQueries}
\title{Get the queries for a project and scenario(s) for display}
\usage{
getScenarioQueries(rFileinfo, scenarios, concat = NULL)
}
\arguments{
\item{rFileinfo}{Reactive fileinfo object returned by file browser in the UI.}

\item{scenarios}{List of scenarios.}

\item{concat}{Separator string for concatenating query names.}
}
\description{
Get the queries for a project and scenario(s) for display
}
